requirements.txt
pytest~=8.4.1
selene==2.0.0rc9
allure-pytest
selenium~=4.35.0
allure-python-commons~=2.15.0